Remigrate.py crashed if files to remove haven't existed.

+++ imgag_project/remigrate.py
@@ -1,4 +1,4 @@
-from os.path import join
+from os.path import join, exists
 from os import remove
 from subprocess import Popen
 import glob
@@ -15,12 +15,15 @@ if __name__ == '__main__':
     MIGRATE_ARGS = [PYTHON, MANAGE_PY, "migrate"]
 
     print("Removing DB...")
-    remove(DB_PATH)
+    if exists(DB_PATH):
+        remove(DB_PATH)
     print("Removing file from media/uploads,media/profile_images and media/categories...")
     for folder_to_delete in UPLOADED_FILE_PATHS:
-        shutil.rmtree(folder_to_delete)
+        if exists(folder_to_delete):
+            shutil.rmtree(folder_to_delete)
     print("Removing migrations...")
-    shutil.rmtree(MIGRATIONS_PATH)
+    if exists(MIGRATIONS_PATH):
+        shutil.rmtree(MIGRATIONS_PATH)
     print("Making migrations for imgag...")
     p = Popen(MAKE_MIGRATE_ARGS)
     p.communicate(input=None)
